Move amd64 HAL into kernel_hal
