--- hal/hal_amd64/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "hal_amd64"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[lib]
-#crate-type = ["cdylib", "static"]
-
-[dependencies]
-bitflags = "2.4.0"
-kernel_api = { path = "../../kernel_api" }
-log = "0.4.20"
