+++ kernel_hal/Cargo.toml
@@ -6,3 +6,6 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+log = "0.4.20"
+bitflags = "2.4.1"
+kernel_api = { path = "../kernel_api" }
